wpt_babel/content-security-policy/securitypolicyviolation/blockeduri-eval.html

<!DOCTYPE html>
<meta content="script-src 'self' 'unsafe-inline'" http-equiv="Content-Security-Policy"/>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<script>
 "use strict";

async_test(function (t) {
  var watcher = new EventWatcher(t, document, 'securitypolicyviolation');
  watcher.wait_for('securitypolicyviolation').then(t.step_func_done(function (e) {
    assert_equals(e.blockedURI, "eval");
    assert_equals(e.lineNumber, 15);
    assert_equals(e.columnNumber, 13);
  }));
  try {
    eval("assert_unreached('eval() should be blocked.");
  } catch (e) {
    assert_equals(e.name, 'EvalError');
  }
}, "Eval violations have a blockedURI of 'eval'");
</script>
